# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "description": "The shape of every non-409 error response.",
        "example": {
          "detail": "invalid credentials"
        },
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KdfParamsOut": {
        "example": {
          "key_length": 32,
          "mem_limit_bytes": 268435456,
          "ops_limit": 3,
          "parallelism": 1,
          "salt": "AAECAwQFBgcICQoLDA0ODw=="
        },
        "properties": {
          "key_length": {
            "default": 32,
            "description": "Derived key length in bytes.",
            "maximum": 64.0,
            "minimum": 16.0,
            "title": "Key Length",
            "type": "integer"
          },
          "mem_limit_bytes": {
            "description": "Argon2id memory cost in bytes. 19 MiB is the OWASP floor; the client ships 256 MiB.",
            "maximum": 4294967296.0,
            "minimum": 19922944.0,
            "title": "Mem Limit Bytes",
            "type": "integer"
          },
          "ops_limit": {
            "description": "Argon2id time cost (iterations).",
            "maximum": 64.0,
            "minimum": 2.0,
            "title": "Ops Limit",
            "type": "integer"
          },
          "parallelism": {
            "default": 1,
            "description": "Argon2id lanes.",
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Parallelism",
            "type": "integer"
          },
          "salt": {
            "description": "Base64 of 16-32 random bytes. Not a secret.",
            "maxLength": 64,
            "minLength": 16,
            "title": "Salt",
            "type": "string"
          }
        },
        "required": [
          "salt",
          "ops_limit",
          "mem_limit_bytes"
        ],
        "title": "KdfParamsOut",
        "type": "object"
      },
      "LoginIn": {
        "example": {
          "auth_secret": "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8=",
          "email": "you@example.com"
        },
        "properties": {
          "auth_secret": {
            "description": "Same derived value used at registration.",
            "maxLength": 64,
            "minLength": 40,
            "title": "Auth Secret",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "totp_code": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional. Supply it to complete MFA in one step instead of via /auth/mfa/verify.",
            "title": "Totp Code"
          }
        },
        "required": [
          "email",
          "auth_secret"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "LogoutOut": {
        "properties": {
          "logged_out": {
            "default": true,
            "title": "Logged Out",
            "type": "boolean"
          }
        },
        "title": "LogoutOut",
        "type": "object"
      },
      "MfaActivateIn": {
        "properties": {
          "code": {
            "description": "A code from the enrolled secret, proving the clock.",
            "maxLength": 10,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "MfaActivateIn",
        "type": "object"
      },
      "MfaActivateOut": {
        "properties": {
          "totp_enabled": {
            "default": true,
            "title": "Totp Enabled",
            "type": "boolean"
          }
        },
        "title": "MfaActivateOut",
        "type": "object"
      },
      "MfaEnrollOut": {
        "properties": {
          "otpauth_uri": {
            "description": "otpauth:// URI for an authenticator app.",
            "title": "Otpauth Uri",
            "type": "string"
          },
          "secret": {
            "description": "Base32 TOTP secret. Inert until /auth/mfa/activate succeeds.",
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret",
          "otpauth_uri"
        ],
        "title": "MfaEnrollOut",
        "type": "object"
      },
      "MfaRequiredOut": {
        "description": "Returned by /auth/login instead of tokens when TOTP is enabled and no\ncode was supplied. Exchange `mfa_token` at /auth/mfa/verify.",
        "properties": {
          "mfa_required": {
            "default": true,
            "title": "Mfa Required",
            "type": "boolean"
          },
          "mfa_token": {
            "description": "Short-lived challenge token. Not an access token.",
            "title": "Mfa Token",
            "type": "string"
          }
        },
        "required": [
          "mfa_token"
        ],
        "title": "MfaRequiredOut",
        "type": "object"
      },
      "MfaVerifyIn": {
        "properties": {
          "code": {
            "description": "Current TOTP code.",
            "maxLength": 10,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          },
          "mfa_token": {
            "title": "Mfa Token",
            "type": "string"
          }
        },
        "required": [
          "mfa_token",
          "code"
        ],
        "title": "MfaVerifyIn",
        "type": "object"
      },
      "RefreshIn": {
        "properties": {
          "refresh_token": {
            "description": "The token from the last login or refresh. Replaying a spent one revokes the family.",
            "maxLength": 256,
            "minLength": 20,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "type": "object"
      },
      "RegisterIn": {
        "example": {
          "auth_secret": "AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8=",
          "email": "you@example.com",
          "kdf": {
            "key_length": 32,
            "mem_limit_bytes": 268435456,
            "ops_limit": 3,
            "parallelism": 1,
            "salt": "AAECAwQFBgcICQoLDA0ODw=="
          }
        },
        "properties": {
          "auth_secret": {
            "description": "Base64 of exactly 32 bytes: BLAKE2b(Argon2id(passphrase, salt), \"zkvault/v1/server-auth\"). Never the passphrase itself.",
            "maxLength": 64,
            "minLength": 40,
            "title": "Auth Secret",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "kdf": {
            "$ref": "#/components/schemas/KdfParamsIn"
          }
        },
        "required": [
          "email",
          "auth_secret",
          "kdf"
        ],
        "title": "RegisterIn",
        "type": "object"
      },
      "RegisterOut": {
        "properties": {
          "user_id": {
            "description": "Opaque account id. Not needed to authenticate.",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "RegisterOut",
        "type": "object"
      },
      "TokenOut": {
        "properties": {
          "access_token": {
            "description": "Bearer token for the Authorization header.",
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Access token lifetime in seconds.",
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "description": "Single use; rotates on every refresh.",
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "zkvault",
      "url": "https://github.com/ahmedazizabbassi/pwd_manager"
    },
...
    "license": {
      "name": "MIT"
    },
    "summary": "Stores encrypted vault blobs. The server cannot decrypt them, cannot search them, and offers no password recovery.",
    "title": "zkvault blind sync",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/kdf-params": {
      "get": {
        "description": "Public KDF parameters for an account, so a new device can derive its keys.\n\nUnknown accounts get deterministic decoy parameters instead of a 404: the\nendpoint must answer before authentication, so a truthful 404 would be a free\naccount-existence oracle.",
        "operationId": "kdf_params_api_v1_auth_kdf_params_get",
        "parameters": [
          {
            "description": "Account email. Unknown addresses get decoy parameters, not a 404.",
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "description": "Account email. Unknown addresses get decoy parameters, not a 404.",
              "title": "Email",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KdfParamsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "summary": "Public KDF parameters for an account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "Returns a token pair, or -- when TOTP is enabled and no `totp_code` was supplied -- an MFA challenge to redeem at `/auth/mfa/verify`. Unknown accounts and wrong secrets are deliberately indistinguishable, in body and in timing.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TokenOut"
                    },
                    {
                      "$ref": "#/components/schemas/MfaRequiredOut"
                    }
                  ],
                  "title": "Response Login Api V1 Auth Login Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown account, wrong secret, or wrong TOTP code."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "summary": "Exchange the derived secret for tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revokes every refresh token for the user.\n\nAccess tokens stay valid until they expire; keeping a revocation list for\nthem would mean a database read per request. The 15-minute TTL is the\ntrade-off, and it is the reason that TTL is short.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke every refresh token for the account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/activate": {
      "post": {
        "operationId": "mfa_activate_api_v1_auth_mfa_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaActivateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaActivateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "No enrolment in progress, or the code did not verify."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finish TOTP enrolment",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "description": "Hands out a TOTP secret; it is inert until /mfa/activate proves the clock.\n\nMFA guards *sync access only*. It cannot gate decryption -- the vault opens\noffline from the passphrase alone, and pretending otherwise would be theatre.",
        "operationId": "mfa_enroll_api_v1_auth_mfa_enroll_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Missing, expired or rejected credentials."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "MFA is already enabled for this account."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start TOTP enrolment",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "mfa_verify_api_v1_auth_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Expired challenge token, or wrong code."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "summary": "Redeem an MFA challenge for tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotating refresh with reuse detection.\n\nPresenting a token that was already spent means either a replay or a stolen\ntoken, and we cannot tell which -- so the whole family is revoked and both\nthe attacker and the legitimate client are forced back through login.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unknown, expired, or already-spent token. Reuse revokes the whole family."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "summary": "Rotate a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Takes the derived `auth_secret` and the public KDF parameters, and stores `Argon2id(auth_secret)`. Nothing sent here can decrypt a vault.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "An account with this email already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Rate limit exceeded. Honour the Retry-After header."
          }
        },
        "summary": "Create an account",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local dev server (make serve)",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}